fix count
this was returning the number of occurrences of the song, not
the number of shows which contain the song, which may not be
the same.  the band frequently reprises a song later in the show.
before this change that would cause the song to be counted twice,
but we want to count shows that contain the song, not the number
of times the song was played.